+++ Tests/Functional/App/config/security_lexik_jwt.yml
