--- Tests/Functional/ForwardCompatTestCaseTrait.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Lexik\Bundle\JWTAuthenticationBundle\Tests\Functional;
-
-/**
- * @internal
- */
-trait ForwardCompatTestCaseTrait
-{
-    protected function tearDown(): void
-    {
-        static::ensureKernelShutdown();
-        static::$kernel = null;
-    }
-
-    protected function setUp(): void
-    {
-        $this->doSetUp();
-    }
-}
