+++ Tests/Functional/SubscribedTokenAuthenticationTest.php
@@ -19,9 +19,9 @@ class SubscribedTokenAuthenticationTest extends CompleteTokenAuthenticationTest
 {
     private static ?object $subscriber = null;
 
-    protected function doSetUp()
+    protected function setUp(): void
     {
-        parent::doSetUp();
+        parent::setUp();
 
         self::$subscriber = static::$kernel->getContainer()->get('lexik_jwt_authentication.test.jwt_event_subscriber');
     }
