--- Tests/Functional/app/config/GenerateTokenCommand/config.yml
@@ -1,7 +1,0 @@
-imports:
-    - { resource: '../base_config.yml' }
-
-lexik_jwt_authentication:
-    secret_key: testing
-    encoder:
-        signature_algorithm: 'HS256'
