+++ Tests/Services/KeyLoader/RawKeyLoaderTest.php
@@ -11,11 +11,11 @@ use Lexik\Bundle\JWTAuthenticationBundle\Services\KeyLoader\RawKeyLoader;
  */
 class RawKeyLoaderTest extends AbstractTestKeyLoader
 {
-    public function doSetUp()
+    protected function setUp(): void
     {
         $this->keyLoader = new RawKeyLoader('private.pem', 'public.pem', 'foobar');
 
-        parent::doSetUp();
+        parent::setUp();
     }
 
     public function testLoadPublicKey()
