+++ composer.json
@@ -51,12 +51,13 @@
         "symfony/translation-contracts": "^1.0|^2.0|^3.0"
     },
     "require-dev": {
+        "phpunit/phpunit": "^9.5",
         "symfony/browser-kit": "^5.4|^6.0",
         "symfony/console": "^5.4|^6.0",
         "symfony/dom-crawler": "^5.4|^6.0",
         "symfony/filesystem": "^5.4|^6.0",
         "symfony/framework-bundle": "^5.4|^6.0",
-        "symfony/phpunit-bridge": "^5.4|^6.0",
+        "symfony/phpunit-bridge": "^6.0",
         "symfony/var-dumper": "^5.4|^6.0",
         "symfony/yaml": "^5.4|^6.0"
     },
@@ -77,10 +78,10 @@
     },
     "scripts": {
         "test": [
-            "vendor/bin/simple-phpunit",
-            "ALGORITHM=HS256 vendor/bin/simple-phpunit",
-            "ENCODER=user_id_claim vendor/bin/simple-phpunit",
-            "PROVIDER=lexik_jwt vendor/bin/simple-phpunit"
+            "vendor/bin/phpunit",
+            "ALGORITHM=HS256 vendor/bin/phpunit",
+            "ENCODER=user_id_claim vendor/bin/phpunit",
+            "PROVIDER=lexik_jwt vendor/bin/phpunit"
         ]
     }
 }
